installer: Set rabbitmq nodename on all rabbitmq hosts.
The existing code for `rabbitmq-server` detection is unfortunately
neither idempotent nor correct.  In only writing out a `nodename` if
rabbitmq is _not_ installed, it is not idempotent; after the first run
of the installer, rabbitmq will now be installed, and the second
`zulip.conf` will thus omit the `nodename`.  While the rabbitmq
configuration files on disk will be left unchanged after the second
run (with `rabbit@localhost` as the nodename), the `zulip.conf` will
not be able to replicate the installation if used elsewhere, as it
will lack the `nodename` setting.

It also fails to set the nodename in cases where a prior step has
installed `rabbitmq-server` already (e.g. `test-install/prepare-base`,
`tools/provision`).  In both of these cases, the lack of `nodename` in
the configuration file will cause later problems if the hostname
changes.  It makes `test-install` unsuitable for testing upgrades via
cloning, for instance.

The potential reason to omit the nodename change if rabbitmq is
already installed is that this results in changing the location where
rabbitmq's data is stored.  The old data is preserved, merely under a
directory with the previous nodename.

For the rare cases where administrators wish to install Zulip on host
with other software deployments, document this potential.  For new
installs, predicate the configuration on if rabbitmq is will be
configured on the host, not its current configuration.

Even though the setting in zulip.conf only ever takes one value, which
is always expected to be set, we must maintain the existence of the
setting for backwards compatibility with installs which predate its
existance, and have data with the hostname as part of the nodename.