export: Write realms active if just-deactiviated.
The command:
```
./manage.py export -r some-realm --deactivate-realm
```
...deactivates the realm before export, to ensure that we get all of
the data, and that it's a fully consistent snapshot.

But it also means that the export includes that the realm is
deactivated, and we preserve that upon import -- which can make for a
very confusing experience upon finishing the migration.

If an export was started with `--deactivate-realm`, override the
exported realm data with `realm.deactivated = False` before export.