restore-backup: Restart go-camo for new passwords.
After running `zulip-puppet-apply` to make sure we have an updated set
of services that should be running, we restart `go-camo` if installed,
to make sure that it picks up the new shared camo secret which we may
have just dropped into place.

`restart-server` does not restart `go-camo`, and it gets the secret
via environment variable read at process startup, so an explicit
service restart is required to pick up the new secret.