test_classes: Delete test data in ZulipTransactionTestCase.tearDown.
Previously, each ZulipTransactionTestCase test was responsible for
deleting its own data, to prevent leakage.  This a hassle for tests,
since it's not always clear what data is created by each test -- for
example, changing the can_send_message_group of a stream can create
additional messages for notification, as well as a RealmAuditLog
entry.

Instead, delete new rows created by each test, by checking if there
are new ids since the test ran. This is a more generic approach; while
it does not catch updates to existing rows, it simplifies cleanup
significantly.

Note that we still keep the verification logic inside tearDown(), as a
double-check that the test didn't leak any data.